fixing table DetailSetupMataPelajaran

+++ database/migrations/2024_02_22_125407_create_detail_setup_mata_pelajarans_table.php
@@ -13,16 +13,18 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('detail_setup_mata_pelajarans', function (Blueprint $table) {
-            $table->id();
-            $table->time('jam_pelajaran');
-            $table->integer('kkm');
-            $table->foreignId('setup_mata_pelajaran_id')->constrained()->onDelete('cascade');
-            $table->foreignId('mata_pelajaran_id')->constrained()->onDelete('cascade');
-            $table->foreignId('pengajar_id')->constrained()->onDelete('cascade');
-            $table->foreignId('tahun_ajaran_id')->constrained('tahun_ajarans')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('mata_pelajarans')) {
+            Schema::create('detail_setup_mata_pelajarans', function (Blueprint $table) {
+                $table->id();
+                $table->time('jam_pelajaran');
+                $table->integer('kkm');
+                $table->foreignId('setup_mata_pelajaran_id')->constrained()->onDelete('cascade');
+                $table->foreignId('mata_pelajaran_id')->constrained()->onDelete('cascade');
+                $table->foreignId('pengajar_id')->constrained()->onDelete('cascade');
+                $table->foreignId('tahun_ajaran_id')->constrained('tahun_ajarans')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
